backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
httpx
python-jose
passlib[bcrypt]
email-validator
pymongo
redis[hiredis]